updated to make compatible to rails 3/redmine 2.1

+++ config/routes.rb
@@ -1,6 +1,15 @@
-ActionController::Routing::Routes.draw do |map|
-  map.connect 'timesheet/index', :controller => 'timesheet', :action => 'index'
-  map.connect 'timesheet/context_menu', :controller => 'timesheet', :action => 'context_menu'
-  map.connect 'timesheet/report.:format', :controller => 'timesheet', :action => 'report'
-  map.connect 'timesheet/reset', :controller => 'timesheet', :action => 'reset', :conditions => { :method => :delete }
+if Rails::VERSION::MAJOR >= 3
+  RedmineApp::Application.routes.draw do
+    match 'timesheet/index' => 'timesheet#index'
+    match 'timesheet/context_menu' => 'timesheet#context_menu'
+    match 'timesheet/report' => 'timesheet#report'
+    match 'timesheet/reset' => 'timesheet#reset', :via => :delete
+  end
+else
+  ActionController::Routing::Routes.draw do |map|
+    map.connect 'timesheet/index', :controller => 'timesheet', :action => 'index'
+    map.connect 'timesheet/context_menu', :controller => 'timesheet', :action => 'context_menu'
+    map.connect 'timesheet/report.:format', :controller => 'timesheet', :action => 'report'
+    map.connect 'timesheet/reset', :controller => 'timesheet', :action => 'reset', :conditions => { :method => :delete }
+  end
 end
